Adding a specific syntax to tell to generalize over hypotheses in induction.
Formerly, generalization over hypotheses was done using "in" which has
also used to select occurrences of the induction term.

This commit adds a special syntax "induction term over hyps". By doing
so, the idea is to provide with a more explicit terminology and to
avoid the confusion between the hypotheses to generalized and the
selection of occurrences in hypotheses which are anyway generalized
(or unselection of clauses, which have occurrences, but which we don't
want to be part of the induction).