Fixes a border case where "contradiction" was unexpectingly failing.
For that, we demand that contradiction calls "elim" systematically
with a non dependent scheme.

Another advantage is that the ability to use dependent elimination of
proofs by default shall not interfere with the code of "contradiction".

Open question: why to use "elim" rather than (non-dependent) "case" to
eliminate from an empty type seems rather arbitrary.