Kernel: Enforcing dependent arity-based let-preserving form of return predicate.
Along the years, it had become standard that the return predicate has
the canonical form "fun Δ (y:I |Δ|) => body" whenever the normalized
form of the arity is "forall Δ, s", with "let"s in Δ not expanded and
|Δ| extracting the variables of Δ which are not for "let"s.

Part of the kernel started to rely on this canonical form (e.g. the
function "type_case_branches") but it was not formally checked that
the predicate had this canonical form (as noticed by Gaëtan at #6749,
March 5).

This commit adds a formal check that it is the case.