New semantics based on unification for "instantiate (n:=t)".
Instead of instantiating the evar referred to by "n" in the current
goal, we unify the designated the evar _applied to its substitution_
with t.

When the substitution is the identity, this amounts to the former
semantics (up to renaming the variables of t in the context of the
goal with their names in the context of the evar).