Experimenting using a record for the substitution of a notation.
It is unfinished...

It could be ok, but this requires a different style. It is much too
verbose as it is (and a lot of changes).

Also, it is done only for CNotation, not for CPatNotation.

We would need some polymorphism of the substitution, so that it is
applicable both to CNotation and CPatNotation, both to constr_expr (in
Constrintern) and to glob_constr (in Notation_ops).