Improve the printing of evar instances with several same names.
The printing skips the display of variables of the evar context whose
instance is a variable of same name as the variable. However, when two
variables have the same name, the renaming done for the context of
evars is in the reverse direction compared to the renaming done for
the instance, so that the printing algorithm gets confused.

To let the printing be informed of how the renaming done in the evar
context and the renaming done in the evar instance connect, we
additionally store the expected renaming of the instance in the evar
context.

Note that this is not a full fix of the printer: each caller to
new_pure_evar should tell how it renames the variables of the instance
and the tactics from tactics.ml still need to do so.