Improve the printing of evar instances with several same names.
The selection of what variables of the instance to print depends on
the strategy using to rename variables: we ensure that the evar knows
the strategy used to print evar instances in detyping.

Note that this is not a full fix: some new_pure_evar calls would need
to be adapted in tactics.ml too.